Add not-a-knot boundary conditions
This commit adds a support for the not-a-knot boundary conditions. API changes:

- Added cubic_spline::boundary_conditions enum type
- Added cubic_spline::natural and cubic_spline::not_a_knot static members as shortcuts
- Added optional bc (boundary conditions) parameter to the constructor

Usage:

    cubic_spline spline(t, x, cubic_spline::not_a_knot);

The default behavior is unchanged (natural). The not-a-knot boundary conditions may produce zeros in the diagonal elements of the tridiagonal system of equations defining splines. So, this commit also updates the tridiagonal solver to avoid numerical issues.